# perf: optimize schema migrations startup performance