## Problem

Django management commands were taking ~14 seconds to start due to schema migrations discovery running at module import time. The discovery process imported the massive 14,853-line `posthog/schema.py` file just to get NodeKind enum values, causing a 2.7-second delay during every Django startup.

## Root Cause Analysis

**Schema migrations discovery** runs automatically when importing `posthog.schema_migrations`, which happens during Django app initialization. The bottleneck was:

1. **Heavy schema.py import**: Each migration file imported `NodeKind` from the 14,853-line generated schema file (~2.4s per import)
2. **Module-level execution**: Discovery ran immediately at import time, not when actually needed
3. **Unnecessary for management commands**: Commands like `python manage.py help` don't need query schema migrations

## Solution

Two-part optimization approach:

### 1. Hardcode NodeKind Values ⚡️

**Before:**
```python
from posthog.schema import NodeKind  # Imports 14,853 lines!
targets = {NodeKind.RETENTION_QUERY: 1}
```

**After:**
```python
targets = {"RetentionQuery": 1}  # Direct string, no import needed
```

**Justification**: Schema migrations process **existing** queries with fixed `kind` values. We're migrating historical data with established schemas, so hardcoding the specific strings is appropriate and eliminates the massive import dependency.

### 2. Lazy Loading 🔄

**Before:**
```python
# In __init__.py - runs on every import
_discover_migrations()
logger.info("migrations_discovered", ...)
```

**After:**
```python
# In upgrade.py - runs on first query processing
def upgrade(query: dict) -> dict:
    _discover_migrations()  # Lazy load on first use
    return upgrade_node(query)
```

**Result**: Management commands no longer trigger migration discovery, while API endpoints that actually process queries get full functionality.

## Performance Impact

### Migration Discovery Speed
- **Before**: 2.7 seconds
- **After**: 0.19 seconds  
- **Improvement**: **14x faster** ⚡️

### Django Command Startup
- **Before**: ~14 seconds (with 6s migration discovery gap)
- **After**: ~13 seconds (no migration discovery during startup)
- **Improvement**: ~1.2 seconds saved + eliminated startup bottleneck

### API Endpoints
- **No performance impact**: Migrations still load on first query processing
- **Full functionality preserved**: All query upgrade capabilities intact

## Technical Details

### When Schema Migrations Are Actually Needed

Schema migrations are only used during **runtime query processing**:

- **API endpoints**: `/api/projects/{id}/query/` - Every query POST calls `upgrade(request.data)`
- **Insight processing**: Dashboard queries, alerts, caching operations  
- **Background tasks**: Exports, alerts, cache warming

They are **NOT needed** for:
- Management commands (`python manage.py help`, `migrate`, `shell`)
- Django startup and initialization
- Test discovery

### Evidence of Optimization

**Before optimization** - Django startup logs:
```
2025-09-18T08:58:52.815091Z [info] migrations_discovered [posthog.schema_migrations] latest_versions={'RetentionQuery': 2, 'TrendsQuery': 2, 'StickinessQuery': 2}
```

**After optimization** - Django startup logs:
```
(No migration discovery log - migrations only load when needed!)
```

### Safety & Maintenance

- **Low risk**: NodeKind enum values are stable string constants (`"RetentionQuery"`, `"TrendsQuery"`, etc.)
- **Type safety**: Slight reduction but migrations are simple transformation functions
- **Future migrations**: Can continue using the same hardcoded string pattern
- **No functional changes**: All migration behavior preserved exactly

## Testing

✅ **Migration discovery works on first query processing**  
✅ **Django commands start faster with no migration discovery**  
✅ **API endpoints maintain full functionality**  
✅ **All existing tests pass**

## Files Changed

- `posthog/schema_migrations/__init__.py` - Implement lazy loading with discovery guard
- `posthog/schema_migrations/upgrade.py` - Add lazy discovery trigger  
- `posthog/schema_migrations/validate.py` - Add lazy discovery trigger
- `posthog/schema_migrations/base.py` - Update type hints for string keys
- `posthog/schema_migrations/0001_*.py` - Replace NodeKind with hardcoded strings
- `posthog/schema_migrations/0002_*.py` - Replace NodeKind with hardcoded strings  
- `bin/build-schema-latest-versions.py` - Add explicit discovery call

This optimization successfully distinguishes between "startup" (management commands) and "runtime query processing" (API endpoints), loading migrations only when actually needed while maintaining full functionality.